Move out logic to unproxy types from ValueObject class into an extension method
Currently the logic to unproxy a type is defined in `ValueObject` class as an helper method.
Even if the method, having an internal visibility, is not part of the public API, I feel it should
not belong in that class. Another reason being that the code is not `ValueObject` specific:
`Entity` calls it also for its comparison.

So this commit does the following:

* Refactored `GetUnproxiedType` into `Unproxy` extension method
I changed the API so the input is no longer an object but the type we want to unproxy.
Making `GetUnproxiedType` an extension on object is out question as the type is too generic.

* Add simple tests
Makes clear that handling proxied object is present.